fix python2 compatibility issues:
